ENGR00285890 imx6:plugin: update the ROM_API_TABLE_BASE_ADDR for plugin code
i.MX6DQ TO1.5 and i.MX6DL/SOLO change the ROM_API_TABLE_BASE_ADDR
from 0xc0 to 0xc4.Need update the plugin code to sync with this change.

The change as the following for the new TO with i.MX6DQ, i.MX6DL/SOLO:

For i.MX6DQ, if the TO >=1.5, will use the new ROM_API_TABLE_BASE_ADDR=0xc4
For i.MX6DL/S, if the TO >=1.2, will use the new ROM_API_TABLE_BASE_ADDR=0xc4

For the old TO, we will still use the 0xc0 to keep compatible.

Signed-off-by: Jason Liu <r64343@freescale.com>